### 2018-09-16 #956 @whedon
[PRE REVIEW]: Opycleid: A Python package for transformational music theory
**Submitting author:** @AlexPof (<a href="http://orcid.org/0000-0002-3767-0361">Alexandre Popoff</a>)
**Repository:** <a href="https://github.com/AlexPof/opycleid" target ="_blank">https://github.com/AlexPof/opycleid</a>
**Version:** 0.2.5
