### 2018-09-16 b905fbd
Bug fix on paper.md
