### 2018-09-16 d98d068
Bug fixes in paper.*
