+++ paper.bib
@@ -91,11 +91,10 @@ ISSN = {00316016},
 
 @book{Lehman:2018,
   title={Hollywood Harmony: Musical Wonder and the Sound of Cinema},
-  author={Lehman, F. and Lehman, A.P.M.F.},
+  author={Lehman, F.},
   isbn={9780190606435},
   lccn={2017041598},
   series={Oxford music/media series},
-  url={https://books.google.fr/books?id=TfgsswEACAAJ},
   year={2018},
   publisher={Oxford University Press}
 }
@@ -106,7 +105,6 @@ ISSN = {00316016},
   isbn={9780199717477},
   lccn={2010017175},
   series={Oxford Handbooks},
-  url={https://books.google.fr/books?id=NZppAgAAQBAJ},
   year={2010},
   publisher={Oxford University Press}
 }
@@ -117,7 +115,6 @@ ISSN = {00316016},
   isbn={9780199772698},
   lccn={2011008754},
   series={Oxford Studies in Music Theory},
-  url={https://books.google.fr/books?id=rZxZCMRiO9EC},
   year={2012},
   publisher={Oxford University Press, USA}
 }
