### 2018-09-16 #956 @AlexPof
I'm fixing some bugs in the paper.md and paper.bib files.
