### 2018-09-17 #956 @danielskatz
I'm glad to have my musical abilities and affinities recognized, and I will attempt to handle this as the editor.
