### 2018-09-17 #956 @AlexPof
@danielskatz - Thank you for accepting to be the editor for this submission. I'm afraid I'm not familiar enough with the reviewers which are mentionned in the list to make a proper choice. May I perhaps suggest drvinceknight who seems to be familiar with sagemath and Python, and who might be interested by the algebraic and computational aspects ?
