### 2018-09-18 #956 @danielskatz
> I am not affiliated to any university / lab / etc. I work for a private company, and this package is completely unrelated. Can this be a problem for the submission process ?

no, this is fine.

> This package is hosted on PyPi and can be installed via pip, with version number 0.2.5. If the review process highlights some changes to be made, does this mean that the version number will have to be changed and the package updated on PyPi ?

yes, that's likely.
