### 2018-09-23 #956 @danielskatz
@willingc - would you be interested in reviewing this? (mostly putting this here as a recording mechanism, since we will talk in person tomorrow)
