### 2018-09-27 #956 @adammichaelwood
I am happy to be a reviewer on this as well.  I have a music degree and am a python developer and tech writer.
