### 2018-09-27 #981 @danielskatz
@drvinceknight and @adammichaelwood - thanks for agreeing to be reviewers.

Please see the instructions above in the start of this issue, and let me know if you have any problems.
