### 2018-09-28 #956 @willingc
> 👋 @willingc - you mentioned a couple of other people on Monday - Can we ping them here and see if they would be able and willing to review this?

It looks like there are now enough reviewers. In the future, would @mscuthbert or @jefftrevino be willing to consider being a reviewer for music submissions for the Journal of Open Source Software (JOSS)?
