### 2018-09-28 #956 @jefftrevino
@willingc Thanks for thinking of me, but I don't know anything about transformational music theory. I use Python to compose by aggregating instances of notational element classes (as in the Abjad API for Formalized Score Control), but I don't model these kinds of relationships abstractly in my work, except implicitly through generative processes. (Although perhaps I should....)
