### 2018-09-29 #956 @mscuthbert
@jefftrevino I think that @willingc is asking if we’d be willing to review open source projects in music in general and not this particular one. To which I say, “of course!”
