### 2018-10-08 #981 @adammichaelwood
I've reviewed the literature and browsed the code. I have not had the
space to do the complete install and use it. I should be able to do that
 in the next week.

On Sun, Oct 7, 2018 at 7:16 AM Daniel S. Katz <notifications@github.com>
wrote:

> 👋 @adammichaelwood <https://github.com/adammichaelwood> - please let us
> know how you are getting along on this
>
