### 2018-10-09 AlexPof/opycleid#1 @drvinceknight
JOSS Review
> Installation: Does installation proceed as outlined in the documentation?

At present I do not see installation instructions in the documentation. `pip install opycleid` works fine for me so perhaps we can add that to the README and tutorial but also include from source instructions as well?

> A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?

After a recent review of a paper of my own it's been explained to me that `paper.md` needs a specific `# Statement of need` section. I believe this would just correspond to your last two paragraphs.

> Automated tests: Are there automated tests or manual steps described so that the function of the software can be verified?

I do not believe that there are any automated tests. Perhaps your examples (README/docs) could be doctested? Or explicitly described as a way to verify that the software is working correctly?

> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

These are missing (once written they can relate to the point of automated testing).
