### 2018-10-09 #981 @drvinceknight
@AlexPof I have opened https://github.com/AlexPof/opycleid/issues/1 with some initial queries/suggestions. :) :+1:
