### 2018-10-13 191dcb8
JOSS review - READMe and paper modified
