### 2018-10-13 AlexPof/opycleid#1 @AlexPof
I have rewritten the ReadMe file to take account of your remarks.

- Installation instructions and dependencies are now clearly stated. As you noted, pip was the easiest way to install the package.
- The `paper.md` has been modified to include an explicit `# Statement of need section`.
- Manual tests are preferred in this case. I have rewritten and expanded the examples in the ReadMe, and explicitly stated them as tests.
- Contributing guidelines were mentionned in the HTML docs, but not in the ReadMe. This is now the case.

Please do not hesitate to tell me if you have additional commens.
