### 2018-10-13 #981 @AlexPof
@drvinceknight I have made the requested modifications in the repository. I also added some missing DOIs to the `paper.md`.
