### 2018-10-21 #981 @danielskatz
👋 @drvinceknight - note that a number of changes have been made since your initial remarks - perhaps you can check on the response?  And continue your review?
