### 2018-10-22 #981 @drvinceknight
Thanks for the changes @AlexPof, everything looks good to me now. I've left one very minor/small suggestion on https://github.com/AlexPof/opycleid/issues/1 but it's in no way a blocker.
