### 2018-10-24 AlexPof/opycleid#1 @AlexPof
Everything now seems in order, so I'm closing this issue
