### 2018-10-27 0ef2acf
paper.md modified for JOSS
