### 2018-11-06 #981 @danielskatz
@adammichaelwood - What's happened in the last 9 days?
