### 2018-11-19 #981 @danielskatz
@adammichaelwood - What's happened in the last 22 days?
