### 2018-11-28 #981 @AlexPof
@danielskatz I was hoping to be able to cite this in a manuscript submission which is due mid-January, do you think it will possible by then ?
