### 2018-11-28 #981 @danielskatz
@AlexPof - sorry for the delay on this.  I've started looking at it myself from the point of view of a reviewer, and once I'm ok with it, we'll go ahead, which hopefully will be quick.
