### 2018-11-29 c484dd9
FIX: dict.items() is OS-dependent on the order of keys; fixed with sorted
