### 2018-11-30 404eb00
FIX: bug fix on dictionaries enumeration
