### 2018-12-01 #981 @danielskatz
@AlexPof - please add non PyPi build instructions to the README and docs as well.
