### 2018-12-01 #981 @danielskatz
You can also upload this version to PyPi.

After that, you should upload the repo to zenodo or similar, and let me know the DOI of the archived version.
