### 2018-12-02 ba51ea0
ADD: installation details
