### 2018-12-02 #981 @AlexPof
I have added the necessary instructions, uploaded the new version on PyPi, and also uploaded the repo to zenodo. The DOI of the zenodo archived version is 10.5281/zenodo.1857913
