### 2018-12-02 #981 @danielskatz
@AlexPof - your paper is now accepted into JOSS ⚡️🚀💥
